### 2021-12-30 #50 @justamad
Sevivi: A Rendering Tool to Generate Videos With Synchronized Sensor Data 
Submitting Author: Name (@justamad)
Package Name: Sevivi
One-Line Description of Package: A Rendering Tool to Generate Videos With Synchronized Sensor Data
Repository Link: https://github.com/HPI-CH/sevivi
Version submitted: 1.0.3
Version accepted: TBD

---

Sevivi is designed to render plots of sensor data next to a video that was taken synchronously, synchronizing the sensor data precisely to the video. It allows you to investigate why certain patterns occur in your sensor data based on the exact moment in the video.

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.

https://github.com/pyOpenSci/software-review/issues/47

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

Sevivi renders plots of given data next to a given video.

Target audience is researchers working with motion data, e.g., prediction of squat intensity using acceleration data. When these researchers have taken videos of their trials, they might want to see what exactly has produced a certain pattern, helping to differentiate between noise and signal. Sevivi makes this easier, by rendering synchronized videos of both the original video and the sensor data plots synchronously. Synchronization can be done manually, using an IMU on the camera (e.g. https://github.com/DavidGillsjo/VideoIMUCapture-Android/) or by using skeleton data from a tracking software (we tested with an azure kinect).

Our research indicates that no similar python packages or other programs exist.

https://github.com/pyOpenSci/software-review/issues/47

- [X] has an [OSI approved license][OsiApprovedLicense].

We're preparing the paper and looking for suggestions regarding the long-term repository options.

- [X] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
