### 2021-12-30 #50 @enra64
@justamad is taking over from me after the accepted [presubmission inquiry](https://github.com/pyOpenSci/software-review/issues/47)
