### 2022-01-07 #50 @xmnlab
## Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested using CI (e.g. GitHub Actions, Travis-CI, CircleCI, etc).
- [x] **Documentation:** Package has documentation setup (ReadTheDocs, JupyterBook, website,etc.).
- [x] **License:** The package has an OSI accepted license
- [ ] **Archive** Zenodo. (may be post-review): The repository is archived somewhere outside of GitHub. We suggest using Zenodo and created releases for this purpose.
- [ ] **Version** (may be post-review) Does the release version given match the GitHub release (v1.0.0)?

---

## Editor comments

In the presubmission issue I saw that @edgarriba is our first reviewer for this package, right? So, I will search for the seconde reviewer :)

@edgarriba, thank you for volunteering to reviewfor PyOpenSci!

When we have the second reviewer confirmed, I will update this comment with the deadline and any extra information necessary.

---

Reviewers: @edgarriba,
Due date:
