### 2022-01-13 #50 @pmeier
@justamad The repository has no tags. From the commit history I identified [52c8bef](https://github.com/HPI-CH/sevivi/commit/52c8bef206e531c797221a08037306c0c5b0ca59) as the commit that changed the version to `1.0.3` in the `pyproject.toml`. I'm going to use it as base for my review.
