### 2022-01-13 HPI-CH/sevivi#34 @pmeier
CLI example in README is broken
pyOpenSci/software-review#50

---

* sevivi version: `1.0.3`
* Python version: `3.9.9`
* Operating System: Linux

### Description

Run the CLI example from the README

### What I Did

```shell
git clone git@github.com:HPI-CH/sevivi.git
cd sevivi/
pip install sevivi
sevivi test-files/test-data-configs/kinect_sync_squatting.toml
```

Running it like that, I was hit with

```
FileNotFoundError: Missing configuration file test-files/test-data-configs/kinect_sync_squatting.toml
```

Replacing `test-files` with `test_files` in the path, I was hit with

```
TypeError: stat: path should be string, bytes, os.PathLike or integer, not NoneType
```
