### 2022-01-22 #50 @pmeier
- [ ] **Function Documentation:** for all user-facing functions
- [ ] **Examples** for all user-facing functions

    Although not required, I think it would be good to link the contribution guide in the README.

- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

    The package includes a `README.rst`, but I guess that should not be a problem? cc @xmnlab

    The extra "readme" in the title can probably be dropped

- [ ] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)

    A [repostatus.org badge](https://www.repostatus.org/) is missing.

- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.

    - Although not required, it think it would be good to have the name explanation
      > Sevivi (SEnsor VIdeo VIsualizer)

      more prominent in the README rather than burying it in the documentation.

- [ ] Any additional setup required (authentication tokens, etc)

- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

    Citation information is missing.

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

    The API documentation is missing completely. I see that the functions have docstrings, but the HTML documentation hosted on ReadTheDocs [does not include them](https://sevivi.readthedocs.io/en/latest/reference.html).

    In the current state I'm unable to asses this. The usage examples offer no explanation of what `sevivi` did for the input data. Given that this is the core of `sevivi`, I feel like this needs significantly more attention. Of the top of my head, the authors could showcase a before-after-comparison highlighting the synchronization. Although we are talking about videos here, you can have GIFs in the documentation. I'm not insisting on this type of presentation. If the authors come up with better visualizations or the like, I happily accept them.

    - The tests emit [one warning](https://github.com/HPI-CH/sevivi/runs/4223506628) for a deprecated `numpy` usage.
    - Other than a total coverage badge in the README, there is no information about the coverage of the individual parts. Consider using a service like [`codecov.io`](https://about.codecov.io/) or at least run `$ coverage report` in CI.

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

    Hard for me to asses without domain knowledge. I'll leave that up to @edgarriba.

Although the option for JOSS publication is checked in https://github.com/pyOpenSci/software-review/issues/50#issue-1091119847, there is no paper to be found in the repository.

---

After installing, I wanted to run the CLI example from the README, but was hit with [multiple errors](https://github.com/HPI-CH/sevivi/issues/34). The Python example is working, but outputs only rather cryptic information. It took me quite some time to realize, the execution generates a `sevivi.avi` file in the root directory. Even with that found, I have no idea what I'm looking at and for me that is a "theme" across the documentation. The next occasion of that is the usage of the acronym "IMU" in the short description in the README without explaining it. For a lay person like me, the documentation was no real help to understand what is going on, so I stopped my review of the functionality rather quickly.

The authors claim that the
> Target audience is researchers working with motion data
in https://github.com/pyOpenSci/software-review/issues/50#issue-1091119847, but I still feel they will have a steep learning curve ahead of them in the current state of `sevivi`. Maybe @edgarriba can comment on that more.

As  of commit [52c8bef](https://github.com/HPI-CH/sevivi/commit/52c8bef206e531c797221a08037306c0c5b0ca59), I can't recommend `sevivi` for publication for all the reasons mentioned above. For me the most pressing ones are:

- Broken example in the README
- Missing API documentation
- Missing explanation of the examples

I'm happy to do another round if my comments are addressed.
