### 2022-01-24 #50 @xmnlab
> The package includes a README.rst, but I guess that should not be a problem? cc @xmnlab

sure thing, that works fine. thanks!

and thank you for your review!
