### 2022-01-26 #50 @edgarriba
The package includes all the following forms of documentation: https://github.com/kornia/kornia#from-source-with-symbolic-links

  * I suggest to add in the readme too how to install to develop. Reference:  https://github.com/kornia/kornia#from-source
  * It's sometimes convenient to add the link to the docs under the project description in the top-right side of the github page. Github allows to directly add links there. It's easier to find the docs.
- [ ] **Function Documentation:** for all user-facing functions
  - no API reference in the `latest` docs.
- [x] **Examples** for all user-facing functions
  - would be nice to find some example scripts in the repo
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [ ] Any additional setup required (authentication tokens, etc)
  - not applicable
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages
  - not applicable

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
  - I agree with @pmeier about coverage. Codecov outputs a very detailed report of what line of code basically have been tests that helps quite a lot to debug of find potential issues in the code.

#### For packages co-submitting to JOSS

No paper ?

---

- example scripts missing (sent a quick PR)
- development environment missing. I wanted to run the tests in local and I had to create my virtual environment, install `poetry`, etc. Would be great if you can provide a `devel` environment to quick-start the development so that contributions don't have to spend time setting up anything. It can be done through docker images or just scripted venv (as we do [here](https://github.com/kornia/kornia/blob/master/setup_dev_env.sh) in kornia).
- the repo is too heavy already I guess because you include the testing data in the root. I suggest to externalize the testing data in a separated repo or cloud service as artifacts. I suggest reading on that: https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-large-files-on-github#removing-a-file-that-was-added-in-an-earlier-commit (and becareful when applying :)
- Regarding the tests, I couldn't find any unit test that prove the correct synchronization.

Agree with @pmeier about temporally rejecting, the package still needs a small iteration based on our comments. Specially on the documentation side. And also couldn't fin the paper.
