### 2022-04-27 #50 @justamad
Hey @xmnlab, thank you very much for the reminder. Unfortunately, there are no updates so far from our side as we currently lack manpower to push this project forward since the last feedback. We will come back once we have resources to finish the necessary tasks :)

Best regards!
