### 2022-09-01 #50 @lwasser
hi @justamad I am checking in on this review. It has been sitting stale for several months - are there any updates @justamad on responding to the reviews and what your timeline will look like? If you don't think that you can get to the reviews than we may need to close the issue and reopen in the future when there is time.

@edgarriba @pmeier thank you so much for your time and patience in reviewing this package! We would like to add you to [the pyopensci website as reviewers ](https://www.pyopensci.org/contributors/) if you are ok with that.
