### 2022-09-02 #50 @pmeier
@lwasser Is there a separate list for reviewers or just the one for contributors? If it is only the latter, both me and @edgarriba are already on it. I've contributed a package and he reviewed it.
