### 2022-09-02 #50 @lwasser
@pmeier thank you for this question -- i did merge your PR. i know it's confusing now. But what you did was perfect. I have those lists setup using YAML so the goal is once i start working on the website I will be able to add how users have contributed and associate reviewers with packages as well. I just haven't implemented those steps yet so you can't see it on the website. but as long as i have the data, I will be able to do it! Thank you again for supporting pyOpenSci! and let me know if you have any other questions
